Add files via upload
Bitica Token (BDCC), created in 2018, it is based on BEP-20 Smart Technology which is owned by Block Beats Company Registered in Estonia (the first country to legalise Cryptocurreny). Since then, numerous cryptocurrencies & tokens have been created. These are frequently called bit coins, as a blend of BDCC coin alternative. BDCC and its derivatives use decentralized control as opposed to centralized electronic money/central banking systems. The decentralized control is related to the use of bitcoin's blockchain transaction database in the role of a distributed. At Bitica, we believe in financial freedom and economic independence, which is possible through BDCC coin.